requirements.txt
python-telegram-bot==13.7
openai==0.27.0
validators==0.18.2
